refactor: EmotionRecordService 대규모 리팩토링 - 서비스 분할 및 음성파일 저장 개선
- 1000줄짜리 EmotionRecordService를 기능별로 4개 서비스로 분할:
  * EmotionRecordCreationService: 감정 기록 생성 관련
  * EmotionRecordQueryService: 감정 기록 조회 관련
  * EmotionRecordManagementService: 감정 기록 수정/삭제 관련
  * EmotionInsightService: 감정 인사이트 관련

- 음성파일 저장 문제 해결:
  * 저장 경로를 /tmp/melog_audio로 변경 (권한 문제 해결)
  * 파일 저장 실패 시 상세한 에러 메시지 제공
  * 디렉토리 쓰기 권한 확인 로직 추가

- 보안 설정:
  * application-dev.yml을 .gitignore에 추가하여 커밋 방지
  * .DS_Store 파일들 git 추적에서 제거

- 코드 가독성 및 유지보수성 대폭 향상
- 각 서비스는 300줄 이하로 관리 가능한 크기
