+++ .gitignore
@@ -49,3 +49,4 @@ out/
 !.github/workflows/*.yml
 .DS_Store
 **/.DS_Store
+src/main/resources/application-dev.yml
