--- src/main/resources/application-dev.yml
@@ -1,38 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:postgresql://postgres/melog?currentSchema=melog
-    username: melog
-    password: melog
-    driver-class-name: org.postgresql.Driver
-    
-
-  jpa:
-    hibernate:
-      ddl-auto: create-drop   # 기존 테이블을 삭제하고 새로 생성
-    properties:
-      hibernate:
-        default_schema: melog
-
-    show-sql: true        # 콘솔에 SQL 출력
-    properties:
-      hibernate:
-        format_sql: true  # SQL 예쁘게 출력
-
-  flyway:
-    schemas: melog
-    default-schema: melog
-    baseline-on-migrate: true
-
-# Clova 개발용 환경변수 설정
-CLOVA_SPEECH_URL: "https://naveropenapi.apigw.ntruss.com"
-CLOVA_SPEECH_CLIENT_ID: "dev-client-id"
-CLOVA_SPEECH_CLIENT_SECRET: "dev-client-secret"
-CLOVA_SPEECH_STT_ENDPOINT: "/recog/v1/stt"
-CLOVA_SPEECH_DEFAULT_LANG: "Kor"
-CLOVA_SPEECH_MAX_DURATION: 60
-CLOVA_SPEECH_SUPPORTED_FORMATS: "mp3,aac,ac3,ogg,flac,wav"
-CLOVA_SPEECH_TIMEOUT_MS: 10000
-CLOVA_STUDIO_BASE_URL: "https://clovastudio.stream.ntruss.com"
-CLOVA_STUDIO_API_KEY: "dev-api-key"
-CLOVA_STUDIO_MODEL: "HCX-005"
-CLOVA_STUDIO_TIMEOUT_MS: 8000
